Fix ingestion failing on unalive roots (#78)
* indexer: inline `maybeGetRootCid`

The function name was misleading: It doesn't maybe get a root cid, it always gets one. It's `getRootCid()` which does the maybe part, so if anything, that could be renamed. This function is a wrapper around `getRootCid()` that merely wraps the error message. For now that's not enough to warant its own function.

* indexer: sqlite needs the value to be `null` not `undefined`

* add passing test

* rename test helper

It's a mock implementation