fix: avoid double slash in piece retrieval URLs
Before this change, when the provider's URL ended with `/`, we
constructed an URL with double slash in the path, e.g.

    https://polynomial.computer//piece/baga123

Curio does not support such paths and returns 404.

In this commit, I fix the code building the URL to handle this edge case
and construct a correct URL with a single slash, e.g.

    https://polynomial.computer/piece/baga123

Signed-off-by: Miroslav Bajtoš <oss@bajtos.net>